Add gpu build tag to toggle gpu support
This PR adds a new `gpu` build tag to control building GPU support into
the collector. These previously used platforms as the build constraints,
only building on Linux. This build tag setup allows us to toggle GPU
support on and off regardless to enable building the collector without
CGO and to make future efforts like ARM support simpler.
