+++ .github/workflows/presubmit.yml
@@ -30,3 +30,6 @@ jobs:
         shell: bash
       - name: Build and test
         run: make install-tools && make presubmit
+      - name: Build and test GPU support
+        if: matrix.os == "ubuntu-20.04"
+        run: make install-tools && make presubmit_gpu_support
