+++ receiver/dcgmreceiver/factory_gpu_off_test.go
@@ -12,29 +12,32 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-//go:build gpu
-// +build gpu
+//go:build !gpu
+// +build !gpu
 
 package dcgmreceiver
 
 import (
+	"context"
+	"errors"
 	"testing"
 
 	"github.com/stretchr/testify/require"
+	"go.opentelemetry.io/collector/consumer/consumertest"
+	"go.opentelemetry.io/collector/receiver/receivertest"
+
+	"github.com/GoogleCloudPlatform/opentelemetry-operations-collector/internal/collectorerror"
 )
 
-func TestDcgmMetricSetFloat64(t *testing.T) {
-	var metric dcgmMetric
-	metric.setFloat64(23.0)
-	require.Equal(t, metric.asFloat64(), 23.0)
-	metric.setFloat64(43.0)
-	require.Equal(t, metric.asFloat64(), 43.0)
-}
+func TestCreateMetricsReceiverWithGPUSupportOff(t *testing.T) {
+	factory := NewFactory()
+	cfg := factory.CreateDefaultConfig()
+	receiver, err := factory.CreateMetricsReceiver(
+		context.Background(),
+		receivertest.NewNopCreateSettings(),
+		cfg,
+		consumertest.NewNop())
 
-func TestDcgmMetricSetInt64(t *testing.T) {
-	var metric dcgmMetric
-	metric.setInt64(23)
-	require.Equal(t, metric.asInt64(), int64(23))
-	metric.setInt64(43)
-	require.Equal(t, metric.asInt64(), int64(43))
+	require.True(t, errors.Is(err, collectorerror.ErrGPUSupportDisabled))
+	require.Nil(t, receiver)
 }
