+++ receiver/dcgmreceiver/factory_gpu_on_test.go
@@ -1,4 +1,4 @@
-// Copyright 2022 Google LLC
+// Copyright 2023 Google LLC
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
@@ -12,10 +12,10 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-//go:build !linux
-// +build !linux
+//go:build gpu
+// +build gpu
 
-package nvmlreceiver
+package dcgmreceiver
 
 import (
 	"context"
@@ -26,15 +26,18 @@ import (
 	"go.opentelemetry.io/collector/receiver/receivertest"
 )
 
-func TestCreateMetricsReceiverOnWindows(t *testing.T) {
+func TestCreateMetricsReceiverOnLinux(t *testing.T) {
 	factory := NewFactory()
 	cfg := factory.CreateDefaultConfig()
+	receiverConfig := cfg.(*Config)
+
 	receiver, err := factory.CreateMetricsReceiver(
 		context.Background(),
 		receivertest.NewNopCreateSettings(),
-		cfg,
-		consumertest.NewNop())
+		receiverConfig,
+		consumertest.NewNop(),
+	)
 
-	require.Regexp(t, ".*only supported on Linux.*", err)
-	require.Nil(t, receiver)
+	require.NoError(t, err)
+	require.NotNil(t, receiver, "failed to create metrics receiver")
 }
