+++ receiver/nvmlreceiver/factory_gpu_off.go
@@ -12,18 +12,19 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-//go:build !linux
-// +build !linux
+//go:build !gpu
+// +build !gpu
 
 package nvmlreceiver
 
 import (
 	"context"
-	"errors"
 
 	"go.opentelemetry.io/collector/component"
 	"go.opentelemetry.io/collector/consumer"
 	"go.opentelemetry.io/collector/receiver"
+
+	"github.com/GoogleCloudPlatform/opentelemetry-operations-collector/internal/collectorerror"
 )
 
 func createMetricsReceiver(
@@ -32,5 +33,5 @@ func createMetricsReceiver(
 	rConf component.Config,
 	consumer consumer.Metrics,
 ) (receiver.Metrics, error) {
-	return nil, errors.New("NVML receiver is only supported on Linux")
+	return nil, collectorerror.ErrGPUSupportDisabled
 }
