+++ receiver/nvmlreceiver/factory_gpu_on.go
@@ -12,8 +12,8 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-//go:build linux
-// +build linux
+//go:build gpu
+// +build gpu
 
 package nvmlreceiver
